requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
gunicorn==23.0.0
Markdown==3.7
numpy==2.2.3
packaging==24.2
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.1
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
wheel==0.45.1
